requirements.txt
discord.py
requests
dotenv
sqlite3
os